datasets/bodik/pyproject.toml
[project]
name = "queria-bodik"
version = "0.1.0"
description = "BODIK ODCS dataset for queria"
requires-python = ">=3.13"
dependencies = [
    "dlt[ducklake]>=1.17.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]
